Refactor rabbit bootstrap script
